chore(fixes/adjustments): adjustments/fixes to meed dashboard requirements + refactoring
