+++ packages/storybook/cxl-dashboard/data/cxl-dashboard.playbooks.data.json
@@ -211,7 +211,7 @@
     ],
     "cxl_item_type": "playbook",
     "cxl_breadcrumbs": ["breadcrumbs", "only", "three", "of", "five"],
-    "cxl_tags": ["Playbooks","Drafts", "Not Started"],
+    "cxl_tags": ["Playbooks","Drafts", "Not Started", "Team Playbooks Wynter"],
     "cxl_steps_count": 10,
     "cxl_author": "Mr Author",
     "conversionxl_certificate_sales_page": "https:\/\/cxl.com\/institute\/online-course\/ab-testing-mastery\/",
@@ -244,7 +244,7 @@
     ],
     "cxl_item_type": "playbook",
     "cxl_breadcrumbs": ["breadcrumbs", "only", "three", "of", "five"],
-    "cxl_tags": ["Playbooks","Drafts", "Started"],
+    "cxl_tags": ["Playbooks","Drafts", "Started", "Team Playbooks Wynter"],
     "cxl_steps_count": 10,
     "cxl_author": "Mr Author",
     "conversionxl_certificate_sales_page": "https:\/\/cxl.com\/institute\/online-course\/ab-testing-mastery\/",
@@ -278,7 +278,7 @@
     ],
     "cxl_item_type": "playbook",
     "cxl_breadcrumbs": ["breadcrumbs", "only", "three", "of", "five"],
-    "cxl_tags": ["Playbooks","Peer Review", "Started"],
+    "cxl_tags": ["Playbooks","Peer Review", "Started", "Team Playbooks CXL"],
     "cxl_steps_count": 10,
     "cxl_author": "Mr Author",
     "conversionxl_certificate_sales_page": "https:\/\/cxl.com\/institute\/online-course\/ab-testing-mastery\/",
@@ -311,7 +311,7 @@
     ],
     "cxl_item_type": "playbook",
     "cxl_breadcrumbs": ["breadcrumbs", "only", "three", "of", "five"],
-    "cxl_tags": ["Playbooks","Saved", "Not Started"],
+    "cxl_tags": ["Playbooks","Saved", "Not Started", "Team Playbooks CXL"],
     "cxl_steps_count": 10,
     "cxl_author": "Mr Author",
     "conversionxl_certificate_sales_page": "https:\/\/cxl.com\/institute\/online-course\/ab-testing-mastery\/",
