+++ packages/storybook/cxl-dashboard/data/sidebar.data.json
@@ -83,7 +83,7 @@
         "tags": ["Drafts", "Peer Review", "Saved"]
       }
     ],
-    "tags": ["Drafts", "Peer Review", "Saved"]
+    "tags": ["Drafts", "Peer Review", "Saved", "Team Playbooks CXL", "Team Playbooks Wynter"]
   },
   {
     "title": "Review new contributors",
