+++ packages/storybook/cxl-dashboard/partials/render-playbooks-dashboard.js
@@ -1,31 +1,7 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { html } from 'lit-element';
 import DashboardRenderItems from './dashboard-render-items';
-
-const filterByTags = (playbooksData) => {
-  /**
-   * @type Array
-   */
-  const elements = Array.from(document.querySelectorAll('#selected-tags vaadin-button[selected]'));
-
-  if (elements && elements.length) {
-    const tags = elements.map((el) => el.innerText);
-
-    return playbooksData.filter((data) => {
-      let hasTag = false;
-
-      tags.forEach((tag) => {
-        if (data.cxl_tags && data.cxl_tags.indexOf(tag) !== -1) {
-          hasTag = true;
-        }
-      });
-
-      return hasTag;
-    });
-  }
-
-  return playbooksData;
-};
+import filterByTags from '../filter-by-tags';
 
 const RenderPlaybooksDashboard = (miniDegData, playbooksData) => {
   const filteredPlaybooks = filterByTags(playbooksData);
